Viết đăng ký đăng nhập và thêm các seeder

+++ database/migrations/2024_11_16_075815_create_membership_fees_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
 
             $table->string('fee_type');
-            $table->decimal('amount');
+            $table->integer('amount');
             $table->boolean('is_open');
             $table->timestamps();
         });
